+++ public/data/Jungmin/commentData.json
@@ -1,19 +1,19 @@
 [
   {
     "id": 1,
-    "userName": "wecode",
+    "user": "wecode",
     "content": " Welcome to world best coding bootcamp!",
     "isLiked": true
   },
   {
     "id": 2,
-    "userName": "joonsikyang",
+    "user": "joonsikyang",
     "content": "Hi there.",
     "isLiked": false
   },
   {
     "id": 3,
-    "userName": "jayPark",
+    "user": "jayPark",
     "content": "Hey.",
     "isLiked": false
   }
