+++ src/pages/Jungmin/Login/InputContainer/Andline/Andline.js
@@ -1,0 +1,15 @@
+import React from 'react';
+import './Andline.scss';
+class Andline extends React.Component {
+  render() {
+    return (
+      <div className="andLine">
+        <div className="line"></div>
+        <div className="and">또는</div>
+        <div className="line"></div>
+      </div>
+    );
+  }
+}
+
+export default Andline;
