+++ src/pages/Jungmin/Login/InputContainer/Andline/Andline.scss
@@ -1,0 +1,18 @@
+.andLine {
+  display: flex;
+  align-items: center;
+  margin: 12px auto 41px;
+  color: rgba(var(--f52, 142, 142, 142), 1);
+
+  .line {
+    height: 1px;
+    width: 98px;
+    background-color: rgba(var(--f52, 142, 142, 142), 1);
+    font-size: 13px;
+    font-weight: 600;
+  }
+
+  .and {
+    margin: 0 18px;
+  }
+}
