+++ src/pages/Jungmin/Login/InputContainer/Button/Button.scss
@@ -1,0 +1,18 @@
+.loginButton {
+  margin: 10px auto;
+  width: 100%;
+  height: 30px;
+  background-color: rgba(var(--d69, 0, 149, 246), 0.3);
+  border: 1px solid transparent;
+  border-radius: 4px;
+  cursor: pointer;
+  color: rgba(var(--eca, 255, 255, 255), 1);
+}
+
+.loginButtonOn {
+  background-color: rgba(var(--d69, 0, 149, 246), 1);
+}
+
+.loginButtonOff {
+  background-color: rgba(var(--d69, 0, 149, 246), 0.3);
+}
