+++ src/pages/Jungmin/Login/InputContainer/InputBox/InputBox.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import './InputBox.scss';
+
+class InputBox extends React.Component {
+  render() {
+    const { name, type, id, placeholder, onchange, goTo } = this.props;
+    return (
+      <form className="inputBox" onSubmit={goTo}>
+        <input
+          className="login"
+          name={name}
+          type={type}
+          id={id}
+          placeholder={placeholder}
+          onChange={onchange}
+        />
+      </form>
+    );
+  }
+}
+
+export default InputBox;
