+++ src/pages/Jungmin/Login/InputContainer/InputBox/InputBox.scss
@@ -1,0 +1,14 @@
+.inputBox {
+  margin: 0 auto 6px;
+  width: 100%;
+
+  .login {
+    width: 268px;
+    height: 38px;
+    padding: 9px 0px 7px 8px;
+    background: rgba(var(--b3f, 250, 250, 250), 1);
+    border: 1px solid rgba(var(--ca6, 219, 219, 219), 1);
+    border-radius: 3px;
+    outline: 5px gray;
+  }
+}
