+++ src/pages/Jungmin/Login/InputContainer/InputContainer.scss
@@ -1,57 +1,6 @@
 .inputContainer {
   align-items: center;
   margin-top: 24px;
-  .inputBox {
-    margin: 0 auto 6px;
-
-    width: 100%;
-    .login {
-      width: 268px;
-      height: 38px;
-      padding: 9px 0px 7px 8px;
-      background: rgba(var(--b3f, 250, 250, 250), 1);
-      border: 1px solid rgba(var(--ca6, 219, 219, 219), 1);
-      border-radius: 3px;
-      outline: 5px gray;
-    }
-  }
-
-  .andLine {
-    display: flex;
-    align-items: center;
-    margin: 12px auto 41px;
-    color: rgba(var(--f52, 142, 142, 142), 1);
-    .line {
-      height: 1px;
-      width: 98px;
-      background-color: rgba(var(--f52, 142, 142, 142), 1);
-      font-size: 13px;
-      font-weight: 600;
-    }
-
-    .and {
-      margin: 0 18px;
-    }
-  }
-
-  .loginButton {
-    margin: 10px auto;
-    width: 100%;
-    height: 30px;
-    background-color: rgba(var(--d69, 0, 149, 246), 0.3);
-    border: 1px solid transparent;
-    border-radius: 4px;
-    cursor: pointer;
-    color: rgba(var(--eca, 255, 255, 255), 1);
-  }
-
-  .loginButtonOn {
-    background-color: rgba(var(--d69, 0, 149, 246), 1);
-  }
-
-  .loginButtonOff {
-    background-color: rgba(var(--d69, 0, 149, 246), 0.3);
-  }
 
   .forgetPw {
     color: #00376b;
