+++ src/pages/Jungmin/Main/Feeds/Feed/Feed.scss
@@ -1,0 +1,38 @@
+.feed {
+  display: flex;
+  flex-direction: column;
+  width: 614px;
+  margin-top: 60px;
+  background-color: white;
+  border: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
+
+  .feedimg {
+    height: 612px;
+    width: 612px;
+  }
+
+  .feedmessage {
+    padding: 5px 0px;
+    align-items: center;
+    display: flex;
+    justify-content: space-between;
+
+    .heart {
+      height: 26px;
+    }
+    .heartButton {
+      outline: 0;
+      border: 0;
+      background-color: transparent;
+    }
+
+    .messageButton button {
+      background: 0 0;
+      border: 0;
+      padding: 8px 14px 8px 8px;
+    }
+    i {
+      font-size: 24px;
+    }
+  }
+}
