+++ src/pages/Jungmin/Main/Feeds/Feed/FeedName/FeedName.scss
@@ -1,0 +1,5 @@
+.feedName {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
