+++ src/pages/Jungmin/Main/Main.js
@@ -1,18 +1,19 @@
 import React from 'react';
 import Nav from '../../../Components/Nav/Nav';
-import Feeds from './Feed/Feeds';
+import Feeds from './Feeds/Feeds';
 import MainAside from './MainAside/MainAside';
 import './Main.scss';
 
 export class MainJungmin extends React.Component {
   render() {
+    const { username } = this.props.history.location.state;
     return (
       <div className="All">
         <div className="mainContainer">
           <div>
-            <Feeds userInfo={this.props.history.location.state} />
+            <Feeds userInfo={username} />
           </div>
-          <MainAside userInfo={this.props.history.location.state} />
+          <MainAside userInfo={username} />
         </div>
       </div>
     );
