+++ src/pages/Jungmin/Main/MainAside/MainAside.scss
@@ -1,57 +1,3 @@
-.personal {
-  padding-top: 5px;
-  width: 293px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  padding-bottom: 3px;
-
-  .personalimg {
-    padding: 3px 12px;
-    border-radius: 100%;
-    height: 40px;
-  }
-
-  .name {
-    font-weight: 600;
-    display: block;
-  }
-  .who {
-    font-size: 14px;
-  }
-
-  .Information {
-    font-size: 12px;
-    color: rgba(var(--f52, 142, 142, 142), 1);
-    font-weight: 500;
-  }
-}
-
-.storyBox {
-  background-color: white;
-  border: 1.5px solid rgba(var(--b6a, 219, 219, 219), 1);
-
-  .storybar {
-    padding: 4px;
-  }
-  span.username {
-    padding: 3px 160px 3px 10px;
-  }
-}
-
-.recommendBox {
-  margin-top: 15px;
-  background-color: white;
-  border: 1.5px solid rgba(var(--b6a, 219, 219, 219), 1);
-
-  .recommendbar {
-    padding: 4px;
-  }
-  span.username {
-    padding: 3px 86px 3px 10px;
-  }
-}
-
 .mainAside {
   display: flex;
   flex-direction: column;
@@ -73,25 +19,33 @@
     border-radius: 100%;
     height: 56px;
   }
-}
 
-.info {
-  color: rgba(var(--edc, 199, 199, 199), 1);
-  font-size: 11px;
-  font-weight: 400;
-  line-height: 13px;
-  width: 293px;
-  margin-top: 15px;
-}
+  .logoutButton {
+    margin-left: 11px;
+    margin-bottom: 6px;
+    background-color: transparent;
+    padding: 6px;
+    border: 1px solid;
+    font-weight: 600;
+    border-radius: 4%;
+  }
 
-// 공통
+  .bold {
+    font-weight: 600;
+  }
 
-.bold {
-  font-weight: 600;
-}
+  .username {
+    color: rgba(var(--f52, 142, 142, 142), 1);
+    font-weight: 500;
+    font-size: 14px;
+  }
 
-.username {
-  color: rgba(var(--f52, 142, 142, 142), 1);
-  font-weight: 500;
-  font-size: 14px;
+  .info {
+    color: rgba(var(--edc, 199, 199, 199), 1);
+    font-size: 11px;
+    font-weight: 400;
+    line-height: 13px;
+    width: 293px;
+    margin-top: 15px;
+  }
 }
