+++ src/pages/Jungmin/Main/MainAside/Personal/Personal.scss
@@ -1,0 +1,28 @@
+.personal {
+  padding-top: 5px;
+  width: 293px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding-bottom: 3px;
+
+  .personalimg {
+    padding: 3px 12px;
+    border-radius: 100%;
+    height: 40px;
+  }
+
+  .name {
+    font-weight: 600;
+    display: block;
+  }
+  .who {
+    font-size: 14px;
+  }
+
+  .Information {
+    font-size: 12px;
+    color: rgba(var(--f52, 142, 142, 142), 1);
+    font-weight: 500;
+  }
+}
