--- src/pages/Jungmin/Main/Post/NewComment.scss
@@ -1,54 +1,0 @@
-.post {
-  word-break: break-all;
-  p {
-    font-weight: 600;
-    display: inline;
-  }
-
-  .forPost {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding-left: 16px;
-    font-size: 14px;
-    padding-top: 2px;
-
-    #heart {
-      height: 13px;
-      padding: 0 10px 0 16px;
-    }
-  }
-
-  .commendTime {
-    padding-left: 16px;
-    padding-bottom: 5px;
-    border-bottom: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
-  }
-
-  .username {
-    color: rgba(var(--f52, 142, 142, 142), 1);
-    font-weight: 500;
-    font-size: 14px;
-  }
-
-  .sumit {
-    display: flex;
-    justify-content: space-between;
-  }
-  .textMessage {
-    width: 90%;
-    height: 55px;
-    border: 0;
-    outline: 0;
-    padding-left: 16px;
-  }
-  .postButton {
-    opacity: 0.5;
-    margin: 0;
-    border: 0;
-    outline: 0;
-    color: #0095f6;
-    background-color: white;
-    width: 10%;
-  }
-}
