--- src/pages/Jungmin/Main/Post/Post.scss
@@ -1,67 +1,0 @@
-.post {
-  word-break: break-all;
-  p {
-    font-weight: 600;
-    display: inline;
-  }
-
-  .forPost {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    padding-left: 16px;
-    font-size: 14px;
-    padding-top: 2px;
-
-    button {
-      outline: 0;
-      border: 0;
-      background-color: transparent;
-      margin-right: 16px;
-
-      .heart {
-        height: 13px;
-      }
-    }
-  }
-
-  .commendTime {
-    padding-left: 16px;
-    padding-bottom: 5px;
-    border-bottom: 1px solid rgba(var(--b6a, 219, 219, 219), 1);
-  }
-
-  .username {
-    color: rgba(var(--f52, 142, 142, 142), 1);
-    font-weight: 500;
-    font-size: 14px;
-  }
-
-  .sumit {
-    display: flex;
-    justify-content: space-between;
-  }
-  .textMessage {
-    width: 90%;
-    height: 55px;
-    border: 0;
-    outline: 0;
-    padding-left: 16px;
-  }
-  .postButton {
-    margin: 0;
-    border: 0;
-    outline: 0;
-    color: #0095f6;
-    background-color: white;
-    width: 10%;
-  }
-
-  .postButtonOn {
-    opacity: 1;
-  }
-
-  .postButtonoff {
-    opacity: 0.5;
-  }
-}
